requirements.txt
marker-pdf
openai
python-dotenv
PyPDF2
fitz
pymupdf